# src/IO.Swagger/wwwroot/swagger-original.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "<h2> WILL BE UPDATED!</h2></br><b>Incomplete</b> API for car workshop system (lacks employees and parts management).</br><a href=\"https://app.swaggerhub.com/apis/soft_eng/project/0.1.1\">Swagger project (contains models)</a>.",
    "version" : "0.1.1",
    "title" : "Car Workshop",
    "contact" : {
      "email" : "k.baciejowski@gmail.com"
    }
  },
  "host" : "virtserver.swaggerhub.com",
  "basePath" : "/soft_eng/project/0.1.1",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/car/{carId}/insurance" : {
      "get" : {
        "tags" : [ "cars" ],
        "summary" : "Find insurance by car Id",
        "operationId" : "getInsurance",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "description" : "Id of car assigned to insurance",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Insurance"
              }
            }
          },
          "400" : {
            "description" : "Invalid Id supplied"
          },
          "404" : {
            "description" : "Car not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:cars" ]
        } ]
      },
      "put" : {
        "tags" : [ "cars" ],
        "summary" : "Add/overwrite car insurance",
        "description" : "Adds or overwrites car insurance. Performed action is determined by type of insurance (i.e. if liability insurance exists it will be replaced).",
        "operationId" : "putInsurance",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "description" : "Id of car assigned to insurance",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "Insurance",
          "description" : "New insurance object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Client"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          },
          "404" : {
            "description" : "Car not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:cars" ]
        } ]
      }
    },
...
    "/car/{carId}/profile" : {
      "get" : {
        "tags" : [ "cars" ],
        "summary" : "Find diagnostic profile by car Id",
        "operationId" : "getProfile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "description" : "Id of car assigned to the diagnostic profile",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "schema" : {
              "$ref" : "#/definitions/DiagnosticProfile"
            }
          },
          "400" : {
            "description" : "Invalid Id supplied"
          },
          "404" : {
            "description" : "Car not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "read:cars" ]
        } ]
      },
      "put" : {
        "tags" : [ "cars", "employees" ],
        "summary" : "Overwrite car profile",
        "description" : "Overwrites diagnostic profile of designated car.",
        "operationId" : "setProfile",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "carId",
          "in" : "path",
          "description" : "Id of car assigned to the diagnostic profile",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "in" : "body",
          "name" : "Profile",
          "description" : "New diagnostic profile object",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/DiagnosticProfile"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Validation exception"
          },
          "404" : {
            "description" : "Profile not found"
          }
        },
        "security" : [ {
          "workshop_auth" : [ "write:cars" ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "workshop_auth" : {
      "type" : "oauth2",
      "authorizationUrl" : "https://example.com/oauth/dialog",
      "flow" : "implicit",
      "scopes" : {
        "write:clients" : "modify clients",
        "read:clients" : "read clients",
        "write:cars" : "modify cars",
        "read:cars" : "read cars",
        "write:visits" : "modify visits",
        "read:visits" : "read visits",
        "write:payments" : "modify payments",
        "read:payments" : "read payments",
        "write:employees" : "modify employees",
        "read:employees" : "read employees"
      }
    },
    "api_key" : {
      "type" : "apiKey",
      "name" : "api_key",
      "in" : "header"
    }
  },
  "definitions" : {
...
    "Client" : {
      "type" : "object",
      "required" : [ "clientId", "discount", "name", "phoneNumber", "surname" ],
      "properties" : {
        "clientId" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 152
        },
        "name" : {
          "type" : "string",
          "example" : "Jan"
        },
        "surname" : {
          "type" : "string",
          "example" : "Kowalski"
        },
        "phoneNumber" : {
          "type" : "string",
          "example" : "+48 123 456 789"
        },
        "email" : {
          "type" : "string",
          "example" : "kowalski@example.com"
        },
        "discount" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 5
        }
      },
      "example" : {
        "clientId" : 152,
        "phoneNumber" : "+48 123 456 789",
        "surname" : "Kowalski",
        "name" : "Jan",
        "discount" : 5,
        "email" : "kowalski@example.com"
      }
    },
...
    "Insurance" : {
      "type" : "object",
      "required" : [ "coverage", "dateOfExpiry", "type" ],
      "properties" : {
        "dateOfExpiry" : {
          "type" : "string",
          "format" : "date-time",
          "example" : "2020-12-30T23:59:59.999+01"
        },
        "coverage" : {
          "type" : "integer",
          "format" : "int32",
          "example" : 100000
        },
        "type" : {
          "type" : "string",
          "example" : "liability insurance",
          "enum" : [ "liability insurance", "comprehensive cover" ]
        }
      },
      "example" : {
        "coverage" : 100000,
        "dateOfExpiry" : "2020-12-30T23:59:59.999+01",
        "type" : "liability insurance"
      }
    },
...
    "DiagnosticProfile" : {
      "type" : "object",
      "properties" : {
        "engine" : {
          "type" : "string",
          "example" : "2120.00cm3"
        },
        "body" : {
          "type" : "string",
          "example" : "black"
        },
        "lowVoltage" : {
          "type" : "string",
          "example" : "12V"
        },
        "lighting" : {
          "type" : "string"
        },
        "brakes" : {
          "type" : "string"
        },
        "sensors" : {
          "type" : "string",
          "example" : "parking sensors"
        },
        "miscellaneous" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "conditionig" : {
          "type" : "string",
          "example" : "full"
        }
      },
      "example" : {
        "sensors" : "parking sensors",
        "miscellaneous" : [ "miscellaneous", "miscellaneous" ],
        "engine" : "2120.00cm3",
        "lowVoltage" : "12V",
        "brakes" : "brakes",
        "conditionig" : "full",
        "body" : "black",
        "lighting" : "lighting"
      }
    },
...
  }
}